requirements.txt
langchain-openai
langchain-core
langchain-ollama
streamlit
python-dotenv
langserve
fastapi
uvicorn